RATS/pyproject.toml
[tool.poetry]
name = "rats-backend"
version = "0.0.0"
description = "A backend part of RATS: reborned cats"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^5.0"
psycopg = "^3.1.16"
psycopg-binary = "^3.1.16"
django-environ = "^0.11.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
